working version of comparison
